Fix #2703: Added LocaleUtils and modified all components to use it.
