+++ src/main/java-templates/org/primefaces/component/treetable/TreeTableTemplate.java
@@ -35,6 +35,7 @@ import org.primefaces.event.RowEditEvent;
 import org.primefaces.event.data.SortEvent;
 import org.primefaces.model.SortOrder;
 import org.primefaces.util.ComponentUtils;
+import org.primefaces.util.LocaleUtils;
 import javax.faces.event.BehaviorEvent;
 import org.primefaces.component.api.UIData;
 import org.primefaces.model.filter.ContainsFilterConstraint;
@@ -364,19 +365,7 @@ import org.primefaces.model.filter.StartsWithFilterConstraint;
 
     public Locale resolveDataLocale() {
         FacesContext context = this.getFacesContext();
-        Object userLocale = this.getDataLocale();
-        
-        if(userLocale != null) {
-            if(userLocale instanceof String)
-                return ComponentUtils.toLocale((String) userLocale);
-            else if(userLocale instanceof java.util.Locale)
-                return (java.util.Locale) userLocale;
-            else
-                throw new IllegalArgumentException("Type:" + userLocale.getClass() + " is not a valid locale type for datatable:" + this.getClientId(context));
-        } 
-        else {
-            return context.getViewRoot().getLocale();
-        }
+        return LocaleUtils.resolveLocale(this.getDataLocale(), this.getClientId(context));
     }
 
     public ColumnGroup getColumnGroup(String target) {
